Harden audio tap buffer write and simplify downmix mean
Agent-Logs-Url: https://github.com/Igor-C-Assuncao/IA-notetaking/sessions/49e7999a-0cb4-45e5-bcf8-5bd4cd337292

+++ src-python/audio_capture.py
@@ -459,7 +459,7 @@ class MacosSystemAudioMixer:
             return np.array([], dtype=np.int16)
         # Ensure complete LR pairs and downmix stereo -> mono.
         stereo_f32 = stereo_f32[: (stereo_f32.size // 2) * 2].reshape(-1, 2)
-        mono_f32 = stereo_f32.mean(axis=1, dtype=np.float32)
+        mono_f32 = stereo_f32.mean(axis=1)
         # Resample 48000 Hz -> 16000 Hz (ratio 1:3)
         mono_16k = resample_poly(mono_f32, 1, 3).astype(np.float32)
         # Convert float32 -> int16
